requirements.txt
requests
python-dotenv
pyTelegramBotAPI
requests
python-dotenv
pyTelegramBotAPI
PySocks
